Add files via upload
There might be certain warnings as I have used libraries which will be deprecated in the future. AkshayData is the .csv file which contains the data for the analysis. I have only made a single function classification_model which can be used by anyone to perform K folds cross-validation. The reason for not making multiple functions for each part is as I wanted to perform detailed analysis for my data and it would be best to view this in Jupyter notebook as I have commented and written the instructions based on the Jupyter markdown.